Fix 32bit warning

+++ AgilityBook/src/ARBCommon/ARBMsgDigestSHA256.cpp
@@ -1373,7 +1373,7 @@ std::wstring ARBMsgDigestComputeSHA256(
 			break;
 		if (outSize)
 			*outSize += static_cast<size_t>(bytes);
-		SHA256_Update(&ctx256, buffer, bytes);
+		SHA256_Update(&ctx256, buffer, static_cast<size_t>(bytes));
 	}
 
 	char buf[1024];
